LoongArch: simplify various JIRLs with "jr rj" and "ret"
The pseudo-instruction "jr" is supported since the very beginning, and
since we have bumped binutils requirement to 2.41 in
commit cb7954c4c2f46e5d ("LoongArch: Add minuimum binutils required
version"), support for "ret" is also guaranteed.

So, we can simplify every "jirl" usage except for those call sites both
indirect and non-tail, and remove the compatibility macro for "ret".
(BTW this is one of the justifications for the toolchain support.)

Signed-off-by: WANG Xuerui <git@xen0n.name>